Remove "bottom" state GroupExpanderOverlay
the original bottom state "when" condition is wrong, and it's just used
as a fallback state. So we just use the qml default state feature, and
move the "bottom" state value to the GroupExpanderOverlay itself. When
no state in states satisfies, it will automatically revert to the
default value.

REVIEW: 123364