[Task Manager ToolTipDelegate] Let positioner size itself automatically
Setting a Row's or Column's size to childrenRect size is just asking for trouble.

Also instead of messing with the "y" position of items within a Row, make spacing of its
parent bigger which is probably what this code wanted to do in the first place.

While at it also get rid of the "_s" property and in preparation for right-to-left layout
replace "x" by anchors.left.

Differential Revision: https://phabricator.kde.org/D2451