[Task Manager] Add media controls in context menu
This adds common media controls to the context menu of task manager entries.

When a player does not have a window opened (running in background) and it
has been pinned as a launcher, controls are still offered, alleviating the
need for a system tray icon.

In addition, a "Restore" option is added to bring to the front the player
window in this case, if supported by the player.

Also, the "Close" option is replaced by a "Quit" player option if there is no player window.

Differential Revision: https://phabricator.kde.org/D2019