Task Manager: Make the icon size configurable for task buttons in a vertical task manager
Summary: This improves the rendering when the task manager is oriented vertically. Before this change the taskbar had to be made too wide so that the task buttons show both icon and text. With the present change, the preferred icon size can be chosen via task bar configuration dialog so that text is shown also in narrower vertical task bars.

Test Plan:
1. Place the task bar vertically
2. Adjust the task bar width e.g. so that the tray icons appear in rows of three
3. Open some windows a Dolphin window so that there is at least one task open in the task bar

Before: The task buttons show only icons that stretch through the most of the button width. There is no text in the task button

After:
1. Task Manager contect menu > Task Manager Settings... >
    Lower the icon size to Small Medium
    Hit OK
2. The task buttons are showing a bit smaller icons and there is some text shown there too.

Reviewers: davidedmundson, hein, broulik

Reviewed By: hein

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4550