[TaskManager] For context menu placement use accurate enum on right edge
The contextMenuComponent used TopPosedLeftAlignedPopup for its
placement property if it was on the right edge. Until now this had no
influence because of the flawed space correction in plasma-framework.

Since this correctional behavior will get improved, the context menu needs
to use the correct enum for its placement from now on.

Also works correctly with unpatched plasma-frameworks.

Reviewers: #plasma, hein

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D4868