[Task Manager] Add option to toggle grouping on middle click
This adds an option for what to do on middle click which will toggle grouping
of window of a specific application.

This can already be done in the context menu by (un)checking "Allow this program
to be grouped" in "Advanced" but this adds a nice shortcut for this.

In contrast to said context menu entry the option in the ComboBox is shown regardless
of the grouping mode as otherwise it might unintentionally change your setting to
a (seemingly) non-deterministic value if you disable grouping.

BUG: 323567
FIXED-IN: 5.10.0

Differential Revision: https://phabricator.kde.org/D5065