[applets/taskmanager] Match default tooltip spacing in custom tooltip
Summary:
The Task Manager widget's custom tooltips use level 5 Headings for the subtitles, which is
pointless since that size is the same size as a Label with with a taller line height,
which is overridden anyway.

This patch makes it just use basic Labels to be consistent with the default tooltip.

Test Plan:
Task Manager launcher item tooltip has same spacing between title and subtitle as a panel
widget's tooltip

Reviewers: #vdg, #plasma, cblack

Reviewed By: #vdg, #plasma, cblack

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D29766